rename autoSlideshowContinuousSpeed
