+++ packages/lib/src/settings/options/autoSlideshowContinuousSpeed.js
